requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
jsonschema>=4.17
fredapi>=0.5
python-dotenv>=1.0
mkdocs-material>=9.5
pymdown-extensions>=10.0
